fix: Rename replication rs files (#675)
Rename the replication server and client files, move them into serparate
mods to improve code organization. This PR has no code/feature changes
