+++ src/bootstrap/mod.rs
@@ -1,12 +1,1 @@
-pub mod error;
-pub mod replication_rpc_client;
-pub mod service;
-
-#[cfg(test)]
-mod replication_client_test;
-
-#[cfg(test)]
-pub mod replication_test_utils;
-
-pub use error::BootstrapError;
-pub use service::ReplicatorBootstrap;
+pub mod replication;
