+++ src/bootstrap/replication/mod.rs
@@ -1,0 +1,9 @@
+pub mod error;
+pub mod rpc_client;
+pub mod service;
+
+#[cfg(test)]
+mod client_test;
+
+#[cfg(test)]
+pub mod test_utils;
