+++ src/lib.rs
@@ -8,7 +8,6 @@ pub mod mempool;
 pub mod network;
 pub mod node;
 pub mod perf;
-pub mod replication;
 pub mod storage;
 pub mod utils;
 pub mod version;
