+++ src/network/mod.rs
@@ -1,6 +1,7 @@
 pub mod admin_server;
 pub mod gossip;
 pub mod http_server;
+pub mod replication;
 pub mod rpc_extensions;
 pub mod server;
